+++ scripts/write-cjs-entry.js
@@ -1,4 +1,8 @@
-"use strict";
+const fs = require('fs')
+const path = require('path')
+
+const outputPath = path.join(process.cwd(), 'dist/lib/index.js')
+const contents = `"use strict";
 
 const BookingSelector = require("./BookingSelector.js").default;
 
@@ -6,3 +10,6 @@ module.exports = BookingSelector;
 module.exports.BookingSelector = BookingSelector;
 module.exports.default = BookingSelector;
 module.exports.__esModule = true;
+`
+
+fs.writeFileSync(outputPath, contents)